Fix bad humanoid state check
I assumed Humanoid death replicated client->server through internal
RemoteEvents. It doesn't! It's actually replicated through the physics
sender.

Even though the timing and order for this is unknowable to scripts, and
the physics sender operates on different jobs on different timelines
from property replication, I could replace it with a network ownership
check... But we also want to change this in the future so it *does* work
through RemoteEvents or property replication. There was an earlier
change that implemented this that was later abandoned, but I'd like to
pick it up and release it. This would fix a bug where killing a Humanoid
welded to another Humanoid kills them both.

Settled on a general check that the source of the RemoteEvent is the
player that owns this character.