Scale joint friction relative to gravity and part mass
Assuming linear forces due to gravity are the main actor on ragdolls.

To maintain a similar "feel" for various avatars in different gravity
simply linearly scale

In reality the torque due to gravity on the joint is also relative to
the offset from the center of mass to the joint and should factor in the
entire limb, but I think that's a more pose-dependent. This seems to
work for now.